Fix style invalidation bug of logical combinations inside :has().
A sibling change doesn't trigger style invalidation properly when
a :has() contains a logical combination containing a sibling
relationship:

<style>
  .subject { color: green }
  .subject:has(:is(.sibling + .child)) { color: red }
</style>
<div id="subject" class="subject">
  <div id="sibling" class="sibling"></div>
  <div class="child">Must be green</div>
</div>
<script>
  document.body.offsetLeft;
  sibling.classList.remove("sibling");
  document.body.offsetLeft;
  // The text must be green but red
</script>

CheckPseudoHasArgumentContext extracts sibling relationship
information from adjacent combinators inside ':has()' and provides
the information to SelectorChecker so that SelectorChecker marks
following flags:
- AncestorsOrAncestorSiblingsAffectedByHas
- SiblingsAffectedByHas

But the extraction logic haven't considered the adjacent combinators
or sibling relationships in the sub-selector of a logical combination
inside ':has()'. (e.g. :has(:is(.a + .b)), :has(:not(:nth-child(2))))

This CL fixes the bug by extracting sibling relationship information
from a sub-selector of :is(), :where() or :not() inside :has().

Bug: 1448531
Change-Id: Ifd5b0db33d357c530a5b8e467f2a7838e86f9528