Undo CSSTransitionDiscrete interpolation test changes
Now that we are planning to add a new opt-in syntax for
CSSTransitionDiscrete, the test changes which I made for
CSSTransitionDiscrete will not eventually stick. Since these tests are
interop23 tracked, the other browsers would like them to be updated
ASAP.

I thought about also disabling the CSSTransitionDiscrete flag, which
would make this patch a lot cleaner, but that would break demos and
stuff people are working on so I decided against it.

This was asked for here:
https://github.com/web-platform-tests/wpt/issues/39871

CSSTransitionDiscrete opt in syntax:
https://github.com/w3c/csswg-drafts/issues/8857

Change-Id: If28b7e9fc4fc0d6dc851e35007fd7614a9b497a0