Shared Storage: Add integration tests for writing from fetch
We add browser and web tests for our implementation of writing to
shared storage from `fetch(<url>, {sharedStorageWritable: true})`.

This is part of a project to add support to writing to shared storage
from response headers. See
https://github.com/WICG/shared-storage/pull/70 for more information
and https://crrev.com/c/4444068 for a working prototype for `fetch()`.

Fuchsia-Binary-Size: Size increase is unavoidable.
Bug:1434529,1218540
Change-Id: Idac762328d6439fbad2fd60b2415c6eefb6421da