Update the common ancestor finding logic for mouseup with mousedown
The existing function doesn't not work for cases like when mousedown
and mouseup were happened in the same hierarchy but the mousedown target
was a different interactive element than the mouseup target. This
could be easily accomplished by setting `pointer-events: none` or
`display: none` to the `active` pseudo-class of the mousedown target.

This patch fixes this by when the logic tries to find the event target
tree of the mousedown target, it continues building up (by recursively looking
up the parents) the tree instead of stopping at the first interactive element.

Differential Revision: https://phabricator.services.mozilla.com/D178844

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1792110
gecko-commit: 63bffcdc61e8a919349315d958358b5e95d1f0d7
gecko-reviewers: edgar