Get currentcolor to work properly for color attribute for colormix
Color is a special case for CSS as "currentColor" is equivalent to
"inherit." This presents a problem for color-mix as it can be
self-referential, i.e:

#foo {
  color: color-mix(in srgb, currentColor, green);
}

Will look to itself to resolve the currentColor, as opposed to its
parent.

The simplest way to resolve this is by walking up the tree in Color::ApplyValue in longhands_custom until a color is found. Ideally
this work would get factored into the StyleBuilderConverter, but
state.ParentNode() is protected, so all the logic needs to happen here
unless we want to refactor all of that.

In a future CL I'll address the issue of _nested_ currentColors in
this scenario, as well as add a test to WPT that addresses that.

Bug: 1448813
Change-Id: I96458be6008ca30d375c921d7aa8b3b3fe90b1b9