[scroll-start] Apply scroll-start after layout
scroll-start[1] should set the initial/default scroll position of a
scroller. Many different things may cause a scroller's size/layout
to change as a page loads which can affect how scroll-start should
be applied. So, this patch applies scroll-start on each frame until
the scroller has been explicitly scrolled, i.e. by the user or via a
programmatic scroll. After an explicit scroll, scroll-start is no
longer applied.

We propagate scroll-start from the document element to the viewport
so that scroll-start can be applied to the page. We save an
explicitly_scrolled_ bit on the element so that scroll-start
does not prevent saved (explicit) scrolls from surviving changes
to display and overflow.

[1] https://drafts.csswg.org/css-scroll-snap-2/#scroll-start

Bug: 1439807
Change-Id: I5655f8e74b85fe45038f700091065fa11eeac0f8