Close other <details> elements linked by name attribute in tree order.
When <details> elements are linked by their common name attribute values
into an exclusive accordion, and one member of the group is closed, this
closes the linked <details> elements in tree order rather than insertion
order.  This reverts much of 34b2758acd53c3ab7ff32bf418334792ad13e523.
The code is simpler but slower, and it matches typical web platform
patterns more closely (even if those patterns may not be ideal in this
case).

This change is based on discussion in
https://github.com/whatwg/html/pull/9400/files#r1221692104

Bug: 1444057
Change-Id: I5eed9e25ff7698e23738e53445c64c0fc3a1b035